Add asset.sha1 column
First step in switching to content-based asset storage.  Once these are
filled in manually in production step 2 will make this not null and
expose it.

I went back and forth between sha1 and sha256 but in the end sha1 is
just so much faster, and switching at some point won't be too hard.