Studies basic slot inclusion and evolution
Synchronized schema

This provides the most basic "study"-like ability to re-use data.  The
effective permissions at this point, however, are inconsistent: the
volume session list will include the inclusions depending on the
consumer permissions, but further requests will fall back to the
producer permissions.  In practice this is not a problem if permissions
are consistent.

Internally this is a little problematic in that record/container models
end up with the wrong volume.  In fact, this may be the right thing to
do given the intended permission semantics.  In this case, just adding
additional volume= arguments on read-only operations (and joining
through volume_inclusion) would be the right way to proceed from here.
A little care must be taken with edit operations, however.

Fixes assembla-280