Separate js and html versions of pages
By default, only send back a js page (angular boot) with no content.
Add a noscript meta-refresh to the non-js version, as well as a noscript
link.  Add a cannonical link on non-js versions to default.  Explicitly
mark all links as js or non-js.

Some links remain default that should be nojs, but then the refresh
kicks in.  Essentially we are optimizing for the js version of the site,
and falling back (at some additional cost) to the non-js version, but
trying to stay there once fallen back.

The biggest user-visible change (and most annoying part) was the token
handling, which was previously a mess.  It's better now, but uses
the registration page for password resets.  Oh, well.